+++ packages/s2-core/src/cell/merged-cell.ts
@@ -14,12 +14,16 @@ import { drawObjectText } from '@/utils/text';
 export class MergedCell extends DataCell {
   public cells: S2CellType[];
 
+  public isPartiallyVisible: boolean;
+
   public constructor(
     spreadsheet: SpreadSheet,
     cells: S2CellType[],
     meta?: ViewMeta,
+    isPartiallyVisible = true, // 合并的单元格只有部分可见。为了方便 Diff 操作，故新增此属性
   ) {
     super(meta, spreadsheet, cells);
+    this.isPartiallyVisible = isPartiallyVisible;
   }
 
   handleRestOptions(...[cells]: [S2CellType[]]) {
@@ -35,9 +39,7 @@ export class MergedCell extends DataCell {
   protected initCell() {
     // TODO：1、条件格式支持； 2、交互态扩展； 3、合并后的单元格文字布局及文字内容（目前参考Excel合并后只保留第一个单元格子的数据）
     this.drawBackgroundShape();
-    // this.drawStateShapes();
     this.drawTextShape();
-    // this.update();
   }
 
   /**