+++ packages/s2-core/src/common/interface/basic.ts
@@ -318,6 +318,7 @@ export interface MergedCellInfo {
 export type TempMergedCell = {
   cells: S2CellType[];
   viewMeta: ViewMeta;
+  isPartiallyVisible?: boolean;
 };
 
 export type FilterDataItemCallback = (