+++ packages/s2-react/__tests__/unit/utils/options-spec.ts
@@ -34,6 +34,7 @@ describe('Options Tests', () => {
           horizontal: 1,
           vertical: 1,
         },
+        scrollbarPosition: 'content',
         autoResetSheetStyle: true,
         resize: {
           colCellHorizontal: true,
@@ -56,6 +57,7 @@ describe('Options Tests', () => {
       frozenTrailingRowCount: 0,
       frozenTrailingColCount: 0,
       hdAdapter: true,
+      supportCSSTransform: false,
     });
   });